Some misc. refactoring

+++ CvGameCoreDLL/CvMapGenerator.h
@@ -23,7 +23,10 @@ public:
 
 	void addLakes();																			// Exposed to Python
 	DllExport void addRivers();														// Exposed to Python
-	void doRiver(CvPlot* pStartPlot, CardinalDirectionTypes eLastCardinalDirection=NO_CARDINALDIRECTION, CardinalDirectionTypes eOriginalCardinalDirection=NO_CARDINALDIRECTION, int iThisRiverID=-1);	// Exposed to Python
+	void doRiver(CvPlot* pStartPlot,												// Exposed to Python
+			CardinalDirectionTypes eLastCardinalDirection = NO_CARDINALDIRECTION,
+			CardinalDirectionTypes eOriginalCardinalDirection = NO_CARDINALDIRECTION,
+			short iThisRiverID = -1); // advc.opt: was int
 	bool addRiver(CvPlot *pFreshWaterPlot);
 	DllExport void addFeatures();													// Exposed to Python
 	DllExport void addBonuses();													// Exposed to Python
